Dry run refactor

+++ config_yml/config.py
@@ -3,8 +3,8 @@ import os
 import copy
 from pathlib import Path
 import logging
-import yaml
 import shutil
+import yaml
 
 log = logging.getLogger(__name__)
 
@@ -13,16 +13,20 @@ class Config:
     """ Manages a config file that will be generated in the /var/ folder
     """
 
-    def __init__(self, package_name: str, template_path: str, config_file_name: str, dry_run_abs_path: str = None):
+    def __init__(self, package_name: str,
+                 template_path: str,
+                 config_file_name: str,
+                 dry_run: bool = False,
+                 dry_run_abs_path: str = None):
         """_summary_
 
         Args:
             package_name (str): Name of the package owner of the config file
             template_path (str): Path of the template file
             config_path (str): Name of config file (will be placed in home/var/{modulename} folder)
-            dry_run_abs_path (str): None if no dry run selected
-                                    "" if dry run selected, with empsty config file
-                                    "{absolute-path}" if dry run selected, starting with an initial file
+            dry_run (bool): If dry run of not
+            dry_run_abs_path (str): "{absolute-path}" if dry run selected, starting with an initial file copied
+                                    "None", if not initial config
         """
         self._template_path = template_path
         self._config_file_name = config_file_name
@@ -31,16 +35,15 @@ class Config:
         if not os.path.exists(self.homevar):
             os.makedirs(self.homevar)
 
-        if dry_run_abs_path is not None:
-            self.dry_run = True
+        self.dry_run = dry_run
+
+        if dry_run:
             self.homevar = os.path.join(self.homevar, 'dryrun_config')
             if os.path.exists(self.homevar):
                 shutil.rmtree(self.homevar)
             os.makedirs(self.homevar)
-            if  dry_run_abs_path != "":
-                shutil.copy(dry_run_abs_path,  os.path.join(self.homevar, self._config_file_name))           
-        else:
-            self.dry_run = False
+            if  dry_run_abs_path:
+                shutil.copy(dry_run_abs_path,  os.path.join(self.homevar, self._config_file_name))
 
         self.config = {}
 
@@ -48,7 +51,8 @@ class Config:
 
     def __del__(self):
         if self.dry_run:
-            shutil.rmtree(self.homevar)
+            if os.path.exists(self.homevar):
+                shutil.rmtree(self.homevar)
 
     def __getitem__(self, key):
         return self.config.get(key, None)
